# expression: implement vectorized evaluation for `builtinPeriodDiffSig`