<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->

implement vectorized evaluation for builtinPeriodDiffSig , for #12103

### What is changed and how it works?

	$go test -v -benchmem -bench=BenchmarkVectorizedBuiltinTimeFunc -run=BenchmarkVectorizedBuiltinTimeFunc -args "builtinPeriodDiffSig"
	goos: darwin
	goarch: amd64
	pkg: github.com/pingcap/tidb/expression
	BenchmarkVectorizedBuiltinTimeFuncGenerated-4   	1000000000	         0.0116 ns/op	       0 B/op	       0 allocs/op
	BenchmarkVectorizedBuiltinTimeFunc/builtinPeriodDiffSig-VecBuiltinFunc-4         	  107150	     11805 ns/op	       0 B/op	       0 allocs/op
	BenchmarkVectorizedBuiltinTimeFunc/builtinPeriodDiffSig-NonVecBuiltinFunc-4      	   36326	     31862 ns/op	       0 B/op	       0 allocs/op
	PASS
	ok  	github.com/pingcap/tidb/expression	2.970s

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test